jalf/jni corrections and testing

+++ libalf/include/libalf/algorithm_deterministic_inferring_csp_z3.h
@@ -1105,7 +1105,7 @@ class deterministic_inferring_csp_Z3 : public automata_inferring<answer> {
 	{ return ALG_INFERRING_CSP_Z3; };
 
 	virtual void generate_statistics(void) {
-		this->statistics["memory.bytes"] = sizeof(this);
+		this->statistics["memory.bytes"] = (int)sizeof(this);
 	}
 
 	virtual std::basic_string<int32_t> serialize() const
