+++ libalf/include/libalf/algorithm_deterministic_inferring_z3.h
@@ -1134,7 +1134,7 @@ class deterministic_inferring_Z3 : public automata_inferring<answer> {
 	{ return ALG_INFERRING_Z3; };
 
 	virtual void generate_statistics(void) {
-		this->statistics["memory.bytes"] = sizeof(this);
+		this->statistics["memory.bytes"] = (int)sizeof(this);
 	}
 
 	virtual std::basic_string<int32_t> serialize() const
