+++ libalf/include/libalf/algorithm_nfa_inferring_minisat.h
@@ -493,7 +493,7 @@ class nfa_inferring_MiniSat : public automata_inferring<bool> {
 	{ return ALG_INFERRING_MINISAT; };
 
 	virtual void generate_statistics(void) {
-		this->statistics["memory.bytes"] = sizeof(this);
+		this->statistics["memory.bytes"] = (int)sizeof(this);
 	}
 
 	virtual std::basic_string<int32_t> serialize() const
